pyproject.toml
[tool.poetry]
name = "gpt-tui"
version = "0.1.0"
description = "A TUI to interact with ChatGPT"
authors = ["Graham Hough <graham@hough.io>"]
readme = "README.md"
packages = [{include = "gpt_tui"}]

[tool.poetry.dependencies]
python = "^3.10"
textual = {extras = ["dev"], version = "^0.10.1"}
httpx = "^0.23.3"
openapi = "^1.1.0"
python-dotenv = "^0.21.1"
openai = "^0.26.4"
asgiref = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
